refactored Search; allow editing of Search term in List view
