+++ src/styles/components/list-view.scss
@@ -4,8 +4,12 @@
     padding: 0;
   }
   
-  .search-term-input {
+  .search-input {
     width: 100%;
+  }
+
+  .search-button {
+    display: none;
   }
 
   .list-item {
