+++ src/styles/components/search-box.scss
