+++ src/styles/main.scss
@@ -7,6 +7,6 @@
 @import 'variables.scss';
 @import 'mixins.scss';
 @import 'components/global.scss';
-@import 'components/search-view.scss';
+@import 'components/search-box.scss';
 @import 'components/list-view.scss';
 @import 'components/detail-view.scss';
